Move requiremets to the global requirements.txt and fix cleaning

+++ beta/tests/test_sota_checkpoints.py
@@ -440,9 +440,10 @@ Tsc = TestSotaCheckpoints
 @pytest.fixture(autouse=True, scope="class")
 def clean_previous_metrics_dump_dir():
     if os.path.isdir(METRICS_DUMP_PATH):
-        files = [file for file in os.listdir(METRICS_DUMP_PATH) if os.path.isfile(file)]
+        files = [os.path.join(METRICS_DUMP_PATH, file) for file in os.listdir(METRICS_DUMP_PATH)]
         for file in files:
-            os.remove(file)
+            if os.path.isfile(file):
+                os.remove(file)
     yield
